Updated agenda, created style, odt file to latex
There is now a common MathSoc style sheet, which can be used to create
consistent formatting between ALL documents. Policy has the additional
style of sectionbreaks (new page for each section).

Updated the agenda, it is now much more full-fledged. TODO is to make
it more templaty, so it can be coppied and filled in. Seperate repo for
agendas most likely.

Changed the ODT file to a Latex one. These files are all super out of
date, so this isn't accurate. I took the liberty of fixing some typos as
well.