Mathsoc Class
Changing to using a mathsoc class, which uses article as it's base.

Seperating out mathsoc.sty to individual sty files that can be used
seperately. Hopefully we can cut down on teh number of requires in the
class as well.

Made an MOU folder that has a sample orientation mou in it.